fix: insert-mode does not log bindings to keyboard macros
Insert mode logs all key-sequences for repeating. When a key-sequence
invokes another key-sequence, the invoking key-sequence is *not*
logged because the invoked key-sequence will be logged, too. The old
behavior was to log both which led to duplicate command executions,
one for the first key-binding and one for the binding the first was
mapped to. Now only the final mapping, which should call a certain
function, is logged.