Initial commit, snippets and files created/modified by me are included.
Note: python-mode is in semi-forked state of version 6.0.4-6.0.5 and
there's problem if you try to migrate to the new python (e.g. 6.0.10)
since in the new versions functions that used to be in 6.0.4.5
are ain't there anymore and fixing this mess involves hardcore yak-shaving.

Also there's similar problem with Icicles and I would recommend to
leave it alone as long as possible.

Situations like this (python, icicles) are best revolved by adopting
some new library and dropping old one. But I'm yet to see any
viable alternatives to icicles and python.