fixed problem with emacs's standard region becoming active after browse-kill-ring paste action is executed
This problem led me to vim:yank-line-handler function that is used as
yank-handler text property for yanked text by vim but never actually called
by vim. But this function in property is executed by insert-for-yank which is
used by browe-kill-ring.el, and basically this is just plain insertion with
special treatment of some text properties.

The problem is, set-mark gets called in vim's normal mode which virtually never
happens since vim:yank-line-handler is never actually called by vim package.
And this set-mark causes selected region to appear if transient-mark mode is
enabled, which is enabled in my config.