remove eshell completely from my config
I'm not using eshell on a daily basis for more than a year. Moreover,
I tried to use it recently and even though it has some nice features,
it's problems cannot be easily worked around. For instance, it's piping
and redirection mechanisms are braindead and are unlikely to improve.

So I leave this commit message as a warning to not to try resurrect
usage of eshell, it's bad idea.