Speed up `shm-get-nodes' by using caches for created markers
It's wasteful to allocate more that one marker for a single position
in buffer because markers are stored in linked list attached to buffer
and detaching N markers later will cause O(N^2) performance hit thanks
to the linked list. This commit does not solve that completely but it
at least mitigates that.