Fix haskell-indentation do deal better with unbalanced sexps
Namely, when some sexp was not closed somewhere above the current
point, 'haskell-indentation-goto-least-indentation' was jumping
there. But that's not always correct since this sexp may have spanned
several functions. So the correct way to handle this is to see whether
there are anything on 0th column between start of the sexp and current point.