Improve facility to keep track of previous buffer in a window
Don’t allow list of previous buffers to grow indefinitely, deduplicate
it and remove dead buffers when needed. Also correctly handle the
situation when a buffer is killed and previous buffer is shown in a
window - we’d like ‘switch-to-prev-buffer-in-window’ to take us to
previous previous buffer but early implementation got stuck because
current buffer was at the head of the prev buf list.