Big update to abbrev+, now it shouldn’t ever appear in the perf profile
Instead of list of regexes put every string that could trigger an
expansion into a trie (a pair of tries but that’s not essential) and
check character by character backwards (NB expansion triggers are
stored reversed in the trie) whether we match any templates to expand.

This check all of the defined abbrevs simultaneously and has less
false negatives (e.g. Haskell’s #scc was not working because it was
actually inserte as “# scc” thanks to smart operators and space was
an issue because we expected in to be a consecutive word).