Major optimization in vim’s tracking of current command keys
Previous algorithm was doing ‘(setq keys (vconcat keys new-keys))’
which repeatedly copies ‘keys’ and thus is O(N^2). Instead store
‘new-key’ blocks in a linked list each time appending via pointer
to the end.

Also amend how visual insert and append command plug in to the insert
mode. Previously they were executing their block part via
‘vim-normal-mode-hook’ that got triggered within ‘vim:insert-mode-exit’
command. Now ‘vim-insert-mode-exit’ is not a defcmd but a function
with an explicit hook that is run at known time which block insertion
& append populate to do their job. Also tweak how they amend undo
history when they’re finished and make it explicit.

In addition, insert mode got its own command function, which is
mostly what normal mode command function does but without
resets of key state. This should help when in insert mode a defcmd
command is executed and later on the insert sequence is repeated <-
without new changes the characters to repeat are incorrect. See one
of the tests added in this commit.