Fix ‘haskell-indentation-check-34’ test to pass in the presence of ‘Fix haskell-indentation do deal better with unbalanced sexps’ commit
The commit takes expression starting at zero column into account to
counter unbalanced sexps. Usually the thing at zeroth column is the
start of the function we’re in and nothing else occurs on than column
in the region between the function start and current point (except
preprocessor directives which are accounted for by ignoring ^# in
regexp).

But this test violates that assumption for no particular gain - in
Haskell the style is to always indent body, noone writes data
definitions like that so it’s not a really useful test.

Yet after the fix the test results seem bogus - see the indentation
point after ‘bar’ and in the middle of ‘::’ which are nonsensical.
However it’s more or less compatible with what was expected
previously.