Disable annoying binding
This guy hangs up Emacs for a little while since it shows up a hello
document with lots of unicode in various languages so Emacs seems to
load some conversion tables. This process is not interruptable with
C-g which makes it super annoying and it’s of no use to me anyway.