Nixify config. Forgo native compilation, focus on byte-compiling with dumping
Native compilation takes 10 minutes to build, whereas byte-compiling
takes 2 (1 of which is running tests). I tried to improve
parallelization for native compilation in this commit but it doesn’t
work. Try it out in text releases of emacs (> 28.1) and see whether it
works. If it does then maybe consider native compiling again.