No need to consider starting column to find previous indent
Either current column is greater than current line’s indent and then
taking it into account doesn’t filter anything, or it’s less that
indend and then it filters out previous indent that are smaller than
current line’s but greater than colunm (which can be all parent
indents if our column happens to fall on the start of line, i.e. is
equal to 0).