Revert ivy’s 230ae1a018c166addb258e20c51150b9d42edcaa commit to speed up completion by avoiding reallocating whole collection
Original intention of having entries with same car but different cdr
does not hold much sense: the user only sees what’s in car. If there
are two entries with the same car the user will not know which one
will be picked so being careful about index does not help. The car’s
must be unique for whole system to make sense, first and foremost
to the user.