Add recording of flycheck error overlays so that we don’t need to filter flycheck overlays of of all overlays. Make enhanced navigation use overlays since attrap fixes may move remaining error overlays and previous navigation through error messages was jumping to stale positions
NB getting all buffer overlays involves malloc instead of GC so it
makes sense to prefer cons-list of overlays if it’s easy to maintain
it.

Error navigation was querying all buffer overlays each time user was
moving to the next error which would be ok if getting overlays was
GC-only but not great for malloc.

Another reason is that buffer may have loads of unrelated overlays and
filtering them each time we’re jumping to next error is not great
either.

But the real motivation is jumpin to flycheck positions which was
using error messages. Buffer may have slightly changed after we got
the messages so overlays are the proper source of locations in the end.